+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/config/theme/look_and_feel/ConfigurableFlatDarculaLaf.java
@@ -1,8 +1,9 @@
-package org.quiltmc.enigma.gui.config.theme;
+package org.quiltmc.enigma.gui.config.theme.look_and_feel;
 
 import com.formdev.flatlaf.FlatDarculaLaf;
+import org.quiltmc.enigma.gui.config.theme.Theme;
 
-public class ConfigurableFlatDarculaLaf extends AbstractConfigurableFlatDarkLaf {
+public class ConfigurableFlatDarculaLaf extends ConfigurableFlatLaf {
 	public static final String NAME = "Configurable " + FlatDarculaLaf.NAME;
 
 	public static Theme.LookAndFeelColors.Builder createColors() {
@@ -10,12 +11,7 @@ public class ConfigurableFlatDarculaLaf extends AbstractConfigurableFlatDarkLaf 
 		return ConfigurableFlatDarkLaf.createColors();
 	}
 
-	protected ConfigurableFlatDarculaLaf(Theme.LookAndFeelColors colors) {
-		super(NAME, colors);
-	}
-
-	@Override
-	protected Base getBase() {
-		return Base.DARCULA;
+	public ConfigurableFlatDarculaLaf(Theme.LookAndFeelColors colors) {
+		super(NAME, colors, Base.DARCULA);
 	}
 }
