+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/config/theme/look_and_feel/ConfigurableFlatDarkLaf.java
@@ -1,10 +1,11 @@
-package org.quiltmc.enigma.gui.config.theme;
+package org.quiltmc.enigma.gui.config.theme.look_and_feel;
 
 import com.formdev.flatlaf.FlatDarkLaf;
+import org.quiltmc.enigma.gui.config.theme.Theme;
 
 // TODO: see if default colors can be loaded from FlatDarkLaf.properties
 // TODO: investigate 'change theme restart' alert button color
-public class ConfigurableFlatDarkLaf extends AbstractConfigurableFlatDarkLaf {
+public class ConfigurableFlatDarkLaf extends ConfigurableFlatLaf {
 	public static final String NAME = "Configurable " + FlatDarkLaf.NAME;
 
 	public static Theme.LookAndFeelColors.Builder createColors() {
@@ -22,12 +23,7 @@ public class ConfigurableFlatDarkLaf extends AbstractConfigurableFlatDarkLaf {
 			.warningBorder(new Theme.SerializableColor(0xFFAC7920));
 	}
 
-	protected ConfigurableFlatDarkLaf(Theme.LookAndFeelColors colors) {
-		super(NAME, colors);
-	}
-
-	@Override
-	protected Base getBase() {
-		return Base.DARK;
+	public ConfigurableFlatDarkLaf(Theme.LookAndFeelColors colors) {
+		super(NAME, colors, Base.DARK);
 	}
 }
